Phase 1: extract annotations into features/annotations.js
State (annotationMode, annotationKey) is private to the module. Decoupled from
main.js: looks up #markdown-content by id; toggleAnnotationMode no longer calls
syncIOSChrome (its sole caller — the toolbar listener — now does that), and
renderAnnotations re-arms the dblclick handlers itself when annotation mode is
active (folding the old post-render check out of main.js).

Pure refactor. main.js ~2,566 -> ~2,430 lines. Build + lint clean; 297 tests pass.
